data-platform-poc/requirements.txt
kafka-python
snowflake-connector-python
snowflake-sqlalchemy
pandas